Fix caret position
When editing input, the labels are updated and input are rerendered and caret position is set at the end of the value.
This is not right as we can edit label at its beginning and we want the caret to keep its position.
To fix that, we save the caret position before editing labels and restore it after input rendering.